Plot of normal density overlaid on the histogram;
Data is taken from the file Hitters.csv in the datasets  repository. Data is for the baseball leagues. I look at the aggregate runs for all the leagues in this file and create a histogram. then I overlay a plot of the normal density  on the histogram. It appears the data is skewed to the left of the mean.